fix(git-diff): security issues and performance
